refactor(framework) Rename `driver/` to `grid/` and `*driver*.py` to `*grid*.py` (#5129)
