+++ src/py/flwr/server/grid/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2024 Flower Labs GmbH. All Rights Reserved.
+# Copyright 2025 Flower Labs GmbH. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,4 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Flower ServerAppIo service."""
+"""Flower driver SDK."""
+
+
+from .grid import Driver
+from .grpc_grid import GrpcDriver
+from .inmemory_grid import InMemoryDriver
+
+__all__ = [
+    "Driver",
+    "GrpcDriver",
+    "InMemoryDriver",
+]
