+++ src/py/flwr/server/server_app.py
@@ -25,7 +25,7 @@ from flwr.server.strategy import Strategy
 
 from .client_manager import ClientManager
 from .compat import start_driver
-from .driver import Driver
+from .grid import Driver
 from .server import Server
 from .server_config import ServerConfig
 from .typing import ServerAppCallable, ServerFn
@@ -207,7 +207,7 @@ class ServerApp:  # pylint: disable=too-many-instance-attributes
         """
 
         def lifespan_decorator(
-            lifespan_fn: Callable[[Context], Iterator[None]]
+            lifespan_fn: Callable[[Context], Iterator[None]],
         ) -> Callable[[Context], Iterator[None]]:
             """Register the lifespan fn with the ServerApp object."""
